move convertToXBit functions to new MIDIHelper class in MIDI/Utilities
