+++ AudioKit/Common/Internals/AudioKitHelpers.swift
@@ -115,19 +115,6 @@ public func random(in range: ClosedRange<Double>) -> Double {
     return Double(arc4random_uniform(UInt32(precision))) / Double(precision) * width + range.lowerBound
 }
 
-
-public func convertTo16Bit(msb: UInt8, lsb: UInt8) -> UInt16 {
-    return (UInt16(msb) << 8) | UInt16(lsb)
-}
-
-public func convertTo32Bit(msb: UInt8, data1: UInt8, data2: UInt8, lsb: UInt8) -> UInt32 {
-    var value: UInt32 = UInt32(lsb) & 0xFF
-    value |= (UInt32(data2) << 8) & 0xFFFF
-    value |= (UInt32(data1) << 16) & 0xFFFFFF
-    value |= (UInt32(msb) << 24) & 0xFFFFFFFF
-    return value
-}
-
 // MARK: - Normalization Helpers
 
 /// Extension to calculate scaling factors, useful for UI controls
